14 October

+++ src/DailyPractice/FarenheitToCelcius.java
@@ -8,7 +8,7 @@ public class FarenheitToCelcius {
         System.out.println("Enter temperature in°F");
         float faranheit = sc.nextFloat();
         sc.close();
-        float celcius = (float) ((float)(faranheit-32.0f)*(5.0/9.0));
+        float celcius = (float) ((faranheit-32.0f)*(5.0/9.0));
         System.out.println("The temperature in Faranheigt is " +celcius+"°C");
